Migrate Queries and commands to Result Migrate controller to HandleResult

+++ src/Activities.Application/Activities/Commands/DeleteCommand.cs
@@ -3,17 +3,24 @@ using Activities.Infrastructure.Persistance.Context;
 
 namespace Activities.Application.Activities.Commands;
 
-public record class DeleteCommand(Guid Id) : ICommand<bool>;
+public record class DeleteCommand(Guid Id) : ICommand<Result<bool>>;
 
-public sealed class DeleteCommandHandler(ActivityContext context) : ICommandHandler<DeleteCommand, bool>
+public sealed class DeleteCommandHandler(ActivityContext context) : ICommandHandler<DeleteCommand, Result<bool>>
 {
-    public async Task<bool> HandleAsync(DeleteCommand command, CancellationToken cancellationToken)
+    public async Task<Result<bool>> HandleAsync(DeleteCommand command, CancellationToken cancellationToken)
     {
-        var activity = await context.Activities.FindAsync([command.Id], cancellationToken) ??
-                       throw new Exception("Activity not found");
+        var activity = await context.Activities.FindAsync([command.Id], cancellationToken);
+
+        if(activity is null)
+        {
+            return Result<bool>.Failure("Activity not found", 404);
+        }   
+
         context.Activities.Remove(activity);
+
         var result = await context.SaveChangesAsync(cancellationToken);
-        return result > 0;
+
+        return result > 0 ? Result<bool>.Success(true) : Result<bool>.Failure("Failed to delete the activity", 400);
     }
 }
